requirements.txt
apache-airflow==2.3.2
asgiref==3.4.1
beautifulsoup4==4.10.0
boto3==1.21.8
botocore==1.24.8
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
cvxopt==1.3.0
cvxpy==1.1.18
Django==3.2.11
django-crontab==0.7.1
django-dotenv==1.4.2
django-extensions==3.1.5
django-storages==1.12.3
ecos==2.0.10
et-xmlfile==1.1.0
gunicorn==20.1.0
html5lib==1.1
idna==3.3
importlib-metadata==4.11.0
jmespath==0.10.0
joblib==1.1.0
lxml==4.7.1
multitasking==0.0.10
numpy==1.21.5
openpyxl==3.0.9
osqp==0.6.2.post5
pandas==1.3.5
pandas-datareader==0.10.0
psycopg2-binary==2.9.3
pyportfolioopt==1.5.1
python-dateutil==2.8.2
pytz==2021.3
qdldl==0.1.5.post0
requests==2.27.1
s3transfer==0.5.2
scikit-learn==1.0.2
scipy==1.7.3
scs==3.1.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
threadpoolctl==3.1.0
typing_extensions==4.0.1
urllib3==1.26.8
webencodings==0.5.1
yfinance==0.1.70
zipp==3.7.0
Historic-Crypto==1.1.6